Upgrade to Ionic 5 (#47)

+++ src/test/java/de/wasenweg/alfred/unit/ComicConstructionTest.java
@@ -1,17 +1,16 @@
 package de.wasenweg.alfred.unit;
 
 import de.wasenweg.alfred.comics.Comic;
+import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.jupiter.api.Test;
 
 public class ComicConstructionTest {
 
   @Test
   public void constructWithNew() throws Exception {
     final Comic comic = new Comic();
-    assertThat(comic.getCurrentPage()).isEqualTo((short) 0);
+    assertThat(comic.getCurrentPage()).isEqualTo(0);
     assertThat(comic.isRead()).isEqualTo(false);
   }
 
@@ -25,11 +24,11 @@ public class ComicConstructionTest {
         .volume("")
         .number("")
         .position("")
-        .year((short) 2000)
-        .month((short) 5)
+        .year(2000)
+        .month(5)
         .publisher("")
         .build();
-    assertThat(comic.getCurrentPage()).isEqualTo((short) 0);
+    assertThat(comic.getCurrentPage()).isEqualTo(0);
     assertThat(comic.isRead()).isEqualTo(false);
   }
